Made changes to getpair.
Not all available pairs were checked by the find routine.

Further, some architectures have a requirement that pairs
be even registers and others don't.
When a pair is found, by setting the next point to search
from using:

      next = found + 1

a client could call this routine again if the even/odd
requirement is not met the first time.

Lal